+++ tests/testthat/test-checkArgumentsPDTMT.R
@@ -49,6 +49,7 @@ test_that("argument checking for PD-TMT works", {
         stattest = "limma",
         minNbrValidValues = 2,
         minlFC = 0,
+        samSignificance = FALSE,
         nperm = 100,
         volcanoAdjPvalThr = 0.05,
         volcanoLog2FCThr = 1,
@@ -467,6 +468,15 @@ test_that("argument checking for PD-TMT works", {
     expect_error(do.call(.checkArgumentsPDTMT, args),
                  "'minlFC' must be within [0,Inf] (inclusive)",
                  fixed = TRUE)
+
+    ## samSignificance
+    args <- args0
+    args$samSignificance <- "1"
+    expect_error(do.call(.checkArgumentsPDTMT, args),
+                 "'samSignificance' must be of class 'logical'")
+    args$samSignificance <- c(TRUE, FALSE)
+    expect_error(do.call(.checkArgumentsPDTMT, args),
+                 "'samSignificance' must have length 1")
 
     ## nperm
     args <- args0