+++ tests/testthat/test-runMaxQuantAnalysis.R
@@ -42,6 +42,7 @@ test_that("runMaxQuantAnalysis works", {
         stattest = "limma",
         minNbrValidValues = 2,
         minlFC = 0,
+        samSignificance = TRUE,
         nperm = 250,
         volcanoAdjPvalThr = 0.05,
         volcanoLog2FCThr = 1,
@@ -405,6 +406,15 @@ test_that("runMaxQuantAnalysis works", {
     expect_error(do.call(runMaxQuantAnalysis, args),
                  "'minlFC' must be within [0,Inf] (inclusive)",
                  fixed = TRUE)
+
+    ## samSignificance
+    args <- args0
+    args$samSignificance <- "1"
+    expect_error(do.call(runMaxQuantAnalysis, args),
+                 "'samSignificance' must be of class 'logical'")
+    args$samSignificance <- c(TRUE, FALSE)
+    expect_error(do.call(runMaxQuantAnalysis, args),
+                 "'samSignificance' must have length 1")
 
     ## nperm
     args <- args0