+++ NEWS.md
@@ -3,6 +3,7 @@
 * Bugfix in PTM workflow to correctly access comparisons
 * Add function to map UniProtIDs to gene symbols
 * Include STRING networks also for PD-TMT workflow
+* Allow t-test for PD-TMT workflow
 
 # einprot 0.5.18